Convert environment-service.ts from a static class to a component. Split into s3/ssm environment-service.ts that are injected with s3 or ssm.
