+++ src/aws/index.ts
@@ -9,7 +9,8 @@ export * from "./cloud-watch-metrics-ratchet";
 export * from "./dynamo-ratchet";
 export * from "./dynamo-table-ratchet";
 export * from "./ec2-ratchet";
-export * from "./environment-service";
+export * from "./environment-service/s3-environment-service";
+export * from "./environment-service/ssm-environment-service";
 export * from "./hash-spreader";
 export * from "./lambda-event-detector";
 export * from "./lambda-event-type-guards";
